+++ src/components/Calendar/components/Month.tsx
@@ -1,5 +1,5 @@
 import type { FC } from 'react';
-import React from 'react';
+import * as React from 'react';
 import { ThemeContext } from 'styled-components';
 import { LIGHT_THEME } from '#src/components/themes';
 
@@ -21,6 +21,7 @@ export const Month: FC<IMonthCalendarProps> = ({
   onMouseEnter,
   onMouseLeave,
   onClick,
+  highlightSpecialDay,
 }) => {
   const theme = React.useContext(ThemeContext) || LIGHT_THEME;
   const weeks: Array<Date> = [];
@@ -52,6 +53,7 @@ export const Month: FC<IMonthCalendarProps> = ({
           filterDate={filterDate}
           onMouseEnter={handleMouseEnter}
           onClick={handleDayClick}
+          highlightSpecialDay={highlightSpecialDay}
         />
       ))}
     </div>