+++ src/components/SelectTree/SelectTreeNode.tsx
@@ -1,5 +1,5 @@
 import type { ChangeEvent, FC, HTMLAttributes, KeyboardEvent, MouseEvent } from 'react';
-import React from 'react';
+import * as React from 'react';
 import styled from 'styled-components';
 import { typography } from '#src/components/Typography';
 import { Checkbox } from '#src/components/Checkbox';