+++ src/components/form/EditModeField/EditModeField.stories.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import type { ComponentMeta, ComponentStory } from '@storybook/react';
 import { EditModeField } from './index';
 import { withDesign } from 'storybook-addon-designs';