+++ src/components/form/common.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import styled from 'styled-components';
 import { FORM_FIELD_DATA_CONTAINER_ATTRIBUTE } from '#src/components/common/utils/splitDataAttributes';